+++ pact_broker/docker_configuration.rb
@@ -27,6 +27,19 @@ module PactBroker
       space_delimited_string_list_or_default(:webhook_http_method_whitelist)
     end
 
+    def database_configuration
+      if database_url
+        database_configuration_from_url
+      else
+        database_configuration_from_parts
+      end.merge(
+        encoding: 'utf8',
+        sslmode: env_or_nil(:database_sslmode),
+        sql_log_level: (env_or_nil(:sql_log_level) || 'debug').to_sym,
+        log_warn_duration: (env_or_nil(:sql_log_warn_duration) || '5').to_f
+      ).compact
+    end
+
     def base_equality_only_on_content_that_affects_verification_results
       if env_populated?(:base_equality_only_on_content_that_affects_verification_results)
         env(:base_equality_only_on_content_that_affects_verification_results) == 'true'
@@ -59,6 +72,10 @@ module PactBroker
       (env(name) || "").size > 0
     end
 
+    def env_or_nil name
+      env_populated?(name) ? env(name) : nil
+    end
+
     def default property_name
       @default_configuration.send(property_name)
     end
@@ -72,7 +89,6 @@ module PactBroker
     end
 
     class SpaceDelimitedStringList < Array
-
       def initialize list
         super(list)
       end
@@ -97,6 +113,43 @@ module PactBroker
           end
         end.join(' ')
       end
+    end
+
+    private
+
+    def database_url
+      @database_url ||= @env[env_or_nil(:database_url_environment_variable_name) || 'PACT_BROKER_DATABASE_URL']
+    end
+
+    def database_configuration_from_parts
+      database_adapter = env_or_nil(:database_adapter) || 'postgres'
+
+      config = {
+        adapter: database_adapter,
+        user: env(:database_username),
+        password: env(:database_password),
+        host: env(:database_host),
+        database: env(:database_name),
+      }
+
+      if env(:database_port) =~ /^\d+$/
+        config[:port] = env(:database_port).to_i
+      end
+
+      config
+    end
+
+    def database_configuration_from_url
+      uri = URI(database_url)
+
+      {
+        adapter: uri.scheme,
+        user: uri.user,
+        password: uri.password,
+        host: uri.host,
+        database: uri.path.sub(/^\//, ''),
+        port: uri.port&.to_i,
+      }.compact
     end
   end
 end